#13454: namespace tt_metal::distributed; add scaffolding for C++ tests
