+++ tests/tt_metal/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(distributed)
+add_subdirectory(tt_metal)
