+++ tt_metal/distributed/distributed.cpp
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "tt_metal/distributed/distributed.hpp"
+
+namespace tt::tt_metal::distributed {
+
+} // namespace tt::tt_metal::distributed
