+++ tt_metal/distributed/distributed.hpp
@@ -1,0 +1,21 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+namespace tt::tt_metal {
+
+inline namespace v0 {
+
+class Device;
+class Tensor;
+
+} // namespace v0
+
+namespace distributed {
+
+class MeshDevice;
+
+} // namespace tt::tt_metal::distributed
+} // namespace tt::tt_metal
