+++ ttnn/cpp/ttnn/operations/ccl/ccl_host_datastructures.hpp
@@ -8,6 +8,7 @@
 #include "ttnn/cpp/ttnn/tensor/tensor_impl.hpp"
 #include "ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
 #include "ttnn/cpp/ttnn/operations/ccl/ccl_host_types.hpp"
+#include "ttnn/distributed/types.hpp"
 #include <limits>
 
 namespace ttnn {
