+++ ttnn/cpp/ttnn/operations/core/to_layout/to_layout_op.cpp
@@ -4,6 +4,7 @@
 
 #include "to_layout_op.hpp"
 
+#include "ttnn/distributed/types.hpp"
 #include "ttnn/operations/data_movement/pad/pad.hpp"
 #include "ttnn/operations/data_movement/tilize/tilize.hpp"
 #include "ttnn/operations/data_movement/tilize_with_val_padding/tilize_with_val_padding.hpp"
