+++ ttnn/cpp/ttnn/operations/data_movement/move/move.cpp
@@ -8,6 +8,7 @@
 #include "ttnn/common/constants.hpp"
 #include "ttnn/decorators.hpp"
 #include "ttnn/run_operation.hpp"
+#include "ttnn/distributed/api.hpp"
 
 namespace ttnn::operations::data_movement {
 
@@ -19,7 +20,7 @@ bool can_deallocate(const Tensor& input_tensor, bool from_multi_device = false) 
                 return storage.buffer.use_count() == (from_multi_device ? 2 : 1);
             } else if constexpr (std::is_same_v<T, MultiDeviceStorage>) {
                 bool can_dealloc = true;
-                auto input_tensors = get_tensors_from_multi_device_storage(input_tensor);
+                auto input_tensors = distributed::get_tensors_from_multi_device_storage(input_tensor);
                 for (const auto& device_tensor : input_tensors) {
                     can_dealloc &= can_deallocate(device_tensor, true);
                 }
@@ -118,7 +119,7 @@ static inline Tensor move(uint8_t queue_id, const Tensor& input_tensor, const st
 static inline Tensor move_sharded(
     uint8_t queue_id, const Tensor& input_tensor, const std::optional<MemoryConfig>& mem_config) {
     std::vector<Tensor> output_tensors = {Tensor(operation::get_workers_for_op_output({input_tensor}))};
-    bool from_multi_device = is_multi_device_tensor(input_tensor);
+    bool from_multi_device = distributed::is_multi_device_tensor(input_tensor);
     operation::launch_op(
         [from_multi_device, mem_config](
             const std::vector<Tensor>& input_tensors,
