+++ ttnn/cpp/ttnn/tensor/serialization.hpp
@@ -15,8 +15,8 @@ namespace tt_metal {
 
 void dump_tensor(const std::string& file_name, const Tensor& tensor, const std::unordered_map<std::string, std::string>& strategy);
 
-template <typename T>
-Tensor load_tensor(const std::string& file_name, T device = nullptr);
+Tensor load_tensor(const std::string& file_name, Device* device = nullptr);
+Tensor load_tensor(const std::string& file_name, distributed::MeshDevice* device = nullptr);
 
 }  // namespace tt_metalls
